use locks to prevent test races
